Staging -> Main (#54)
* remove leftover model file

* don't create empty circle metadata row

* more sensible default value for DB_HOST in .env.example

* Revert "more sensible default value for DB_HOST in .env.example"

This reverts commit bb564b8cc7cc2dd102ab8a82fe48aea602da13a9.

* initial commit

* adding scoped data for new endpoints

* scoping out all the remaining data endpoint permissions

* add missing jsonresponse

* add v2 to login endpoint

* update getcircle

* replace signature validation with token authentication

* add backwards compatibility for gifts

* move v1 routes back to api.php to make diff cleaner

* after review updates

* remove deprecated upload functions

* add comment about $request->user()

* review cleanup

rename middleware, remove unnecessary imports

* hotfix for profile update

fix backward compatibility with old authentication.

* verify content of signed message when logging in (#55)

* verify content of signed message when logging in

* add simple login route

* update full-circle & manifest endpoint return values

* remove raw sql queries to be postgres compatible

* console command to patch postgres boolean default issue

after converting mysql to postgres the default value doesn't get converted for boolean column types, added console command to reset them to what they are suppose to be.

+++ app/Repositories/ProfileRepository.php
@@ -1,7 +1,11 @@
 <?php
 
 namespace App\Repositories;
+
+use App\Models\Circle;
+use App\Models\Epoch;
 use App\Models\Profile;
+use App\Models\User;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 use Intervention\Image\Facades\Image;
@@ -9,44 +13,50 @@ use Intervention\Image\Facades\Image;
 class ProfileRepository
 {
     protected $model;
+
     public function __construct(Profile $model)
     {
         $this->model = $model;
     }
 
-    public function getProfile($request, $address) {
-        return $this->model->with(['users.circle.protocol','users.teammates','users.histories.epoch'])->byAddress($address)->first();
+    public function getProfile($address, $relationships = [])
+    {
+        return $this->model->with($relationships)->byAddress($address)->first();
     }
 
-    public function saveProfile($request) {
-        $data = $request->only('skills','bio','telegram_username',
-            'discord_username','twitter_username','github_username','medium_username','website');
+    public function saveProfile($request)
+    {
+        $data = $request->only('skills', 'bio', 'telegram_username',
+            'discord_username', 'twitter_username', 'github_username', 'medium_username', 'website');
         $profile = $request->profile;
         $profile->update($data);
-        $profile->load(['users.circle.protocol','users.teammates','users.histories.epoch']);
+        $profile->load(['users.circle.protocol', 'users.teammates', 'users.histories.epoch']);
         return $profile;
     }
 
-    public function uploadProfileAvatar($request, $address) {
-        $profile =  $this->model->byAddress($address)->first();
-        if($profile) {
+    public function uploadProfileAvatar($request)
+    {
+        $profile = $request->user();
+        if ($profile) {
             $file = $request->file('file');
             $image = Image::make($file);
             $height = $image->height();
             $width = $image->width();
 
-            if($width> 240) {
-                $height = $height * 240/$height;
-                $width = $width * 240/$width;
+            if ($width > 240) {
+                $height = $height * 240 / $height;
+                $width = $width * 240 / $width;
             }
 
             $resized = $image
-                ->resize($width, $height, function ($constraint) { $constraint->aspectRatio(); } )
-                ->encode($file->getCLientOriginalExtension(),80);
-            $new_file_name = Str::slug(pathinfo(basename($file->getClientOriginalName()), PATHINFO_FILENAME)).'_'.time().'.'.$file->getCLientOriginalExtension();
+                ->resize($width, $height, function ($constraint) {
+                    $constraint->aspectRatio();
+                })
+                ->encode($file->getCLientOriginalExtension(), 80);
+            $new_file_name = Str::slug(pathinfo(basename($file->getClientOriginalName()), PATHINFO_FILENAME)) . '_' . time() . '.' . $file->getCLientOriginalExtension();
             $ret = Storage::put($new_file_name, $resized);
-            if($ret) {
-                if($profile->avatar && Storage::exists($profile->avatar)) {
+            if ($ret) {
+                if ($profile->avatar && Storage::exists($profile->avatar) && config('app.env') == 'production') {
                     Storage::delete($profile->avatar);
                 }
 
@@ -59,16 +69,17 @@ class ProfileRepository
         return null;
     }
 
-    public function uploadProfileBackground($request, $address) {
-        $profile =  $this->model->byAddress($address)->first();
-        if($profile) {
+    public function uploadProfileBackground($request)
+    {
+        $profile = $request->user();
+        if ($profile) {
             $file = $request->file('file');
             $resized = Image::make($request->file('file'))
-                ->encode($file->getCLientOriginalExtension(),80);
-            $new_file_name = Str::slug(pathinfo(basename($file->getClientOriginalName()), PATHINFO_FILENAME)).'_'.time().'.'.$file->getCLientOriginalExtension();
+                ->encode($file->getCLientOriginalExtension(), 80);
+            $new_file_name = Str::slug(pathinfo(basename($file->getClientOriginalName()), PATHINFO_FILENAME)) . '_' . time() . '.' . $file->getCLientOriginalExtension();
             $ret = Storage::put($new_file_name, $resized);
-            if($ret) {
-                if($profile->background && Storage::exists($profile->background)) {
+            if ($ret) {
+                if ($profile->background && Storage::exists($profile->background) && config('app.env') == 'production') {
                     Storage::delete($profile->background);
                 }
 
@@ -79,5 +90,19 @@ class ProfileRepository
         }
 
         return null;
+    }
+
+    public function getCircleDataWithProfile($profile)
+    {
+        $circle_ids = $profile->circle_ids();
+        $users = User::whereIn('circle_id', $circle_ids)->get();
+        if ($profile->admin_view) {
+            $circles = Circle::all();
+            $epochs = Epoch::isActiveFutureDate()->where('ended', 0)->get();
+        } else {
+            $circles = Circle::whereIn('id', $circle_ids)->get();
+            $epochs = Epoch::isActiveFutureDate()->where('ended', 0)->whereIn('circle_id', $circle_ids)->get();
+        }
+        return compact('users', 'circles', 'epochs');
     }
 }
